chore(service): :hammer: relocate services into a folder
